pyproject.toml
[tool.poetry]
name = "ollama_robusta_actions"
version = "0.0.1"
description = "Ollama actions for Robusta"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
ollama = "0.2.0"
cachetools = "^5.3.0"

[tool.poetry.dev-dependencies]
robusta-cli = "^0.10.14"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
